+++ src/pythonic_fp/homepage/compatible_versions.py
@@ -4,6 +4,25 @@ Known Boring Math compatible versions.
 Coordinated releases
 ====================
 
+2026-01-03 - 3.3.4
+------------------
+
+============== ========================= =======
+Name           PyPI Project              Version
+============== ========================= =======
+Booleans       pythonic-fp-booleans      2.1.0
+Circular Array pythonic-fp-circulararray 6.0.0
+Containers     pythonic-fp-containers    4.0.0
+FP Tools       pythonic-fp-fptools       5.2.0
+Gadgets        pythonic-fp-gadgets       4.0.1
+Iterables      pythonic-fp-iterables     5.1.2
+Numpy          pythonic-fp-numpy         0.1.1
+Queues         pythonic-fp-queues        5.1.0
+Splitends      pythonic-fp-splitends     2.0.1
+============== ========================= =======
+
+----
+
 2025-12-02 - 3.3.3
 ------------------