+++ src/pythonic_fp/homepage/top_level_releases.py
@@ -10,20 +10,21 @@ PyPI and GitHub Name         Python Module               Version Date Released
 pythonic-fp-booleans         pythonic_fp.booleans        2.1.0   2025-12-02
 pythonic-fp-circulararray    pythonic_fp.circulararray   6.0.0   2025-09-26
 pythonic-fp-containers       pythonic_fp.containers      4.0.0   2025-09-28
-pythonic-fp-fptools          pythonic_fp.fptools         5.1.2   2025-09-28
+pythonic-fp-fptools          pythonic_fp.fptools         5.2.0   2026-01-03
 pythonic-fp-gadgets          pythonic_fp.gadgets         4.0.1   2025-12-02
 pythonic-fp-iterables        pythonic_fp.iterables       5.1.2   2025-09-28
 pythonic-fp-queues           pythonic_fp.queues          5.1.0   2025-09-26
 pythonic-fp-numpy            pythonic_fp.numpy           0.1.1   2025-12-02
 pythonic-fp-sentinels [#]_   pythonic_fp.sentinels       2.1.3   2025-08-02
 pythonic-fp-singletons [#]_  pythonic_fp.singletons      1.0.0   2025-09-25
-pythonic-fp-splitends        pythonic_fp.splitends       2.0.0   2025-09-28
+pythonic-fp-splitends        pythonic_fp.splitends       2.0.1   2025-09-28
 ============================ =========================== ======= =============
 
 Remarks
 -------
 
 .. attention::
+
     Maintainer will try to keep these top level releases consistent
     with each other as much as possible.